Rename open editor outline service to reflect that it watches active editors only
