+++ resources/META-INF/plugin_template.xml
@@ -311,7 +311,7 @@
       <implementation-class>io.flutter.editor.WidgetIndentsHighlightingPassFactory</implementation-class>
     </component>
     <component>
-      <implementation-class>io.flutter.editor.outline.OpenEditorOutlineService</implementation-class>
+      <implementation-class>io.flutter.editor.ActiveEditorsOutlineService</implementation-class>
     </component>
   </project-components>
